requirements.txt
certifi==2019.6.16
chardet==3.0.4
idna==2.8
PyYAML==5.1.1
requests==2.22.0
six==1.12.0
slackclient==1.3.1
tableauserverclient==0.8.1
urllib3==1.25.3
websocket-client==0.54.0
